Refactor: move persistance code outside WorkItemStore
