+++ src/aggregator-ruleng/RuleObjects/WorkItemStore.cs
@@ -175,277 +175,19 @@ namespace aggregator.Engine
                     _context.Logger.WriteVerbose($"No save mode specified, assuming {SaveMode.TwoPhases}.");
                     goto case SaveMode.TwoPhases;
                 case SaveMode.Item:
-                    var resultItem = await SaveChanges_ByItem(commit, impersonate, bypassrules, cancellationToken);
+                    var saver = new Persistance.PersistByItem(_context);
+                    var resultItem = await saver.SaveChanges_ByItem(commit, impersonate, bypassrules, cancellationToken);
                     return resultItem;
                 case SaveMode.Batch:
-                    var resultBatch = await SaveChanges_Batch(commit, impersonate, bypassrules, cancellationToken);
+                    var saver2 = new Persistance.PersistBatch(_context);
+                    var resultBatch = await saver2.SaveChanges_Batch(commit, impersonate, bypassrules, cancellationToken);
                     return resultBatch;
                 case SaveMode.TwoPhases:
-                    var resultTwoPhases = await SaveChanges_TwoPhases(commit, impersonate, bypassrules, cancellationToken);
+                    var saver3 = new Persistance.PersistTwoPhases(_context);
+                    var resultTwoPhases = await saver3.SaveChanges_TwoPhases(commit, impersonate, bypassrules, cancellationToken);
                     return resultTwoPhases;
                 default:
                     throw new InvalidOperationException($"Unsupported save mode: {mode}.");
-            }
-        }
-
-        private async Task<(int created, int updated)> SaveChanges_ByItem(bool commit, bool impersonate, bool bypassrules, CancellationToken cancellationToken)
-        {
-            int createdCounter = 0;
-            int updatedCounter = 0;
-
-            var (createdWorkItems, updatedWorkItems, deletedWorkItems, restoredWorkItems) = _context.Tracker.GetChangedWorkItems();
-            foreach (var item in createdWorkItems)
-            {
-                if (commit)
-                {
-                    _context.Logger.WriteInfo($"Creating a {item.WorkItemType} workitem in {item.TeamProject}");
-                    _ = await _clients.WitClient.CreateWorkItemAsync(
-                        item.Changes,
-                        _context.ProjectName,
-                        item.WorkItemType,
-                        bypassRules: impersonate || bypassrules,
-                        cancellationToken: cancellationToken
-                    );
-                }
-                else
-                {
-                    _context.Logger.WriteInfo($"Dry-run mode: should create a {item.WorkItemType} workitem in {item.TeamProject}");
-                }
-
-                createdCounter++;
-            }
-
-            if (commit)
-            {
-                await RestoreAndDelete(restoredWorkItems, deletedWorkItems, cancellationToken);
-            }
-            else if (deletedWorkItems.Any() || restoredWorkItems.Any())
-            {
-                static string FormatIds(WorkItemWrapper[] items) => string.Join(",", items.Select(item => item.Id));
-                var teamProjectName = restoredWorkItems.FirstOrDefault()?.TeamProject ??
-                                      deletedWorkItems.FirstOrDefault()?.TeamProject;
-                _context.Logger.WriteInfo($"Dry-run mode: should restore: {FormatIds(restoredWorkItems)} and delete {FormatIds(deletedWorkItems)} workitems from {teamProjectName}");
-            }
-            updatedCounter += restoredWorkItems.Length + deletedWorkItems.Length;
-
-            foreach (var item in updatedWorkItems.Concat(restoredWorkItems))
-            {
-                if (commit)
-                {
-                    _context.Logger.WriteInfo($"Updating workitem {item.Id}");
-                    _ = await _clients.WitClient.UpdateWorkItemAsync(
-                        item.Changes,
-                        item.Id,
-                        bypassRules: impersonate || bypassrules,
-                        cancellationToken: cancellationToken
-                    );
-                }
-                else
-                {
-                    _context.Logger.WriteInfo($"Dry-run mode: should update workitem {item.Id} in {item.TeamProject}");
-                }
-
-                updatedCounter++;
-            }
-
-            return (createdCounter, updatedCounter);
-        }
-
-        private async Task<(int created, int updated)> SaveChanges_Batch(bool commit, bool impersonate, bool bypassrules, CancellationToken cancellationToken)
-        {
-            // see https://github.com/redarrowlabs/vsts-restapi-samplecode/blob/master/VSTSRestApiSamples/WorkItemTracking/Batch.cs
-            // and https://docs.microsoft.com/en-us/rest/api/vsts/wit/workitembatchupdate?view=vsts-rest-4.1
-            // BUG this code won't work if there is a relation between a new (id<0) work item and an existing one (id>0): it is an API limit
-
-            var (createdWorkItems, updatedWorkItems, deletedWorkItems, restoredWorkItems) = _context.Tracker.GetChangedWorkItems();
-            int createdCounter = createdWorkItems.Length;
-            int updatedCounter = updatedWorkItems.Length + deletedWorkItems.Length + restoredWorkItems.Length;
-
-            List<WitBatchRequest> batchRequests = new List<WitBatchRequest>();
-            foreach (var item in createdWorkItems)
-            {
-                _context.Logger.WriteInfo($"Found a request for a new {item.WorkItemType} workitem in {item.TeamProject}");
-
-                var request = _clients.WitClient.CreateWorkItemBatchRequest(_context.ProjectName,
-                                                                            item.WorkItemType,
-                                                                            item.Changes,
-                                                                            bypassRules: impersonate,
-                                                                            suppressNotifications: false);
-                batchRequests.Add(request);
-            }
-
-            foreach (var item in updatedWorkItems)
-            {
-                _context.Logger.WriteInfo($"Found a request to update workitem {item.Id} in {item.TeamProject}");
-
-                var request = _clients.WitClient.CreateWorkItemBatchRequest(item.Id,
-                                                                            item.Changes,
-                                                                            bypassRules: impersonate || bypassrules,
-                                                                            suppressNotifications: false);
-                batchRequests.Add(request);
-            }
-
-            var converters = new JsonConverter[] { new JsonPatchOperationConverter() };
-            string requestBody = JsonConvert.SerializeObject(batchRequests, Formatting.None, converters);
-            _context.Logger.WriteVerbose(requestBody);
-
-            if (commit)
-            {
-                _ = await ExecuteBatchRequest(batchRequests, cancellationToken);
-                await RestoreAndDelete(restoredWorkItems, deletedWorkItems, cancellationToken);
-            }
-            else
-            {
-                _context.Logger.WriteWarning($"Dry-run mode: no updates sent to Azure DevOps.");
-            }//if
-
-            return (createdCounter, updatedCounter);
-        }
-
-        private static bool IsSuccessStatusCode(int statusCode)
-        {
-            return (statusCode >= 200) && (statusCode <= 299);
-        }
-
-        //TODO no error handling here? SaveChanges_Batch has at least the DryRun support and error handling
-        //TODO Improve complex handling with ReplaceIdAndResetChanges and RemapIdReferences
-        private async Task<(int created, int updated)> SaveChanges_TwoPhases(bool commit, bool impersonate, bool bypassrules, CancellationToken cancellationToken)
-        {
-            // see https://github.com/redarrowlabs/vsts-restapi-samplecode/blob/master/VSTSRestApiSamples/WorkItemTracking/Batch.cs
-            // and https://docs.microsoft.com/en-us/rest/api/vsts/wit/workitembatchupdate?view=vsts-rest-4.1
-            // The workitembatchupdate API has a huge limit:
-            // it fails adding a relation between a new (id<0) work item and an existing one (id>0)
-
-            var (createdWorkItems, updatedWorkItems, deletedWorkItems, restoredWorkItems) = _context.Tracker.GetChangedWorkItems();
-            int createdCounter = createdWorkItems.Length;
-            int updatedCounter = updatedWorkItems.Length + deletedWorkItems.Length + restoredWorkItems.Length;
-
-            //TODO strange handling, better would be a redesign here: Add links as new Objects and do not create changes when they occur but when accessed to Changes property
-            var batchRequests = new List<WitBatchRequest>();
-            foreach (var item in createdWorkItems)
-            {
-                _context.Logger.WriteInfo($"Found a request for a new {item.WorkItemType} workitem in {item.TeamProject}");
-
-                //TODO HACK better something like this: _context.Tracker.NewWorkItems.Where(wi => !wi.Relations.HasAdds(toNewItems: true))
-                var changesWithoutRelation = item.Changes
-                                                 .Where(c => c.Operation != Microsoft.VisualStudio.Services.WebApi.Patch.Operation.Test)
-                                                 // remove relations as we might incour in API failure
-                                                 .Where(c => !string.Equals(c.Path, "/relations/-", StringComparison.Ordinal));
-                var document = new JsonPatchDocument();
-                document.AddRange(changesWithoutRelation);
-
-                var request = _clients.WitClient.CreateWorkItemBatchRequest(_context.ProjectName,
-                                                                            item.WorkItemType,
-                                                                            document,
-                                                                            bypassRules: impersonate || bypassrules,
-                                                                            suppressNotifications: false);
-                batchRequests.Add(request);
-            }
-
-            if (commit)
-            {
-                var batchResponses = await ExecuteBatchRequest(batchRequests, cancellationToken);
-
-                UpdateIdsInRelations(batchResponses);
-
-                await RestoreAndDelete(restoredWorkItems, deletedWorkItems, cancellationToken);
-            }
-            else
-            {
-                _context.Logger.WriteWarning($"Dry-run mode: no updates sent to Azure DevOps.");
-            }
-
-            batchRequests.Clear();
-            var allWorkItems = createdWorkItems.Concat(updatedWorkItems).Concat(restoredWorkItems);
-            foreach (var item in allWorkItems)
-            {
-                var changes = item.Changes
-                                  .Where(c => c.Operation != Microsoft.VisualStudio.Services.WebApi.Patch.Operation.Test);
-                if (!changes.Any())
-                {
-                    continue;
-                }
-                _context.Logger.WriteInfo($"Found a request to update workitem {item.Id} in {_context.ProjectName}");
-                _context.Logger.WriteVerbose(JsonConvert.SerializeObject(item.Changes));
-
-                var request = _clients.WitClient.CreateWorkItemBatchRequest(item.Id,
-                                                                            item.Changes,
-                                                                            bypassRules: impersonate || bypassrules,
-                                                                            suppressNotifications: false);
-                batchRequests.Add(request);
-            }
-
-            if (commit)
-            {
-                _ = await ExecuteBatchRequest(batchRequests, cancellationToken);
-            }
-            else
-            {
-                _context.Logger.WriteWarning($"Dry-run mode: no updates sent to Azure DevOps.");
-            }
-
-            return (createdCounter, updatedCounter);
-        }
-
-        private async Task<IEnumerable<WitBatchResponse>> ExecuteBatchRequest(IList<WitBatchRequest> batchRequests, CancellationToken cancellationToken)
-        {
-            if (!batchRequests.Any()) return Enumerable.Empty<WitBatchResponse>();
-
-            var batchResponses = await _clients.WitClient.ExecuteBatchRequest(batchRequests, cancellationToken: cancellationToken);
-
-            var failedResponses = batchResponses.Where(batchResponse => !IsSuccessStatusCode(batchResponse.Code)).ToList();
-            foreach (var failedResponse in failedResponses)
-            {
-                string stringResponse = JsonConvert.SerializeObject(batchResponses, Formatting.None);
-                _context.Logger.WriteVerbose(stringResponse);
-                _context.Logger.WriteError($"Save failed: {failedResponse.Body}");
-            }
-
-            //TODO should we throw exception?
-#pragma warning disable S125 // Sections of code should not be commented out
-            //if (failedResponses.Any())
-            //{
-            //    throw new InvalidOperationException("Save failed.");
-            //}
-#pragma warning restore S125 // Sections of code should not be commented out
-            return batchResponses;
-        }
-
-
-        private async Task RestoreAndDelete(IEnumerable<WorkItemWrapper> restore, IEnumerable<WorkItemWrapper> delete, CancellationToken cancellationToken = default)
-        {
-            foreach (var item in delete)
-            {
-                _context.Logger.WriteInfo($"Deleting workitem {item.Id} in {item.TeamProject}");
-                _ = await _clients.WitClient.DeleteWorkItemAsync(item.Id, cancellationToken: cancellationToken);
-            }
-
-            foreach (var item in restore)
-            {
-                _context.Logger.WriteInfo($"Restoring workitem {item.Id} in {item.TeamProject}");
-                _ = await _clients.WitClient.RestoreWorkItemAsync(new WorkItemDeleteUpdate() { IsDeleted = false }, item.Id, cancellationToken: cancellationToken);
-            }
-        }
-
-        private void UpdateIdsInRelations(IEnumerable<WitBatchResponse> batchResponses)
-        {
-            var (createdWorkItems, updatedWorkItems, deletedWorkItems, restoredWorkItems) = _context.Tracker.GetChangedWorkItems();
-            var realIds = createdWorkItems
-                                   // the response order matches the request order
-                                   .Zip(batchResponses, (item, response) =>
-                                   {
-                                       int oldId = item.Id;
-                                       var newId = response.ParseBody<WorkItem>().Id.Value;
-
-                                       //TODO oldId should be known by item, and not needed to be passed as parameter
-                                       item.ReplaceIdAndResetChanges(oldId, newId);
-                                       return new { oldId, newId };
-                                   })
-                                   .ToDictionary(kvp => kvp.oldId, kvp => kvp.newId);
-
-            foreach (var item in updatedWorkItems)
-            {
-                item.RemapIdReferences(realIds);
             }
         }
